requirements.txt
pyserial
matplotlib
numpy
pillow